Move page delete view to wagtail.admin.views.pages.delete
