+++ wagtail/admin/views/pages/__init__.py
@@ -1,5 +1,4 @@
 from django.core.exceptions import PermissionDenied
-from django.db import transaction
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect
 from django.template.response import TemplateResponse
@@ -13,6 +12,7 @@ from wagtail.core.models import Page, UserPagePermissionsProxy
 
 from wagtail.admin.views.pages.copy import *  # noqa
 from wagtail.admin.views.pages.create import *  # noqa
+from wagtail.admin.views.pages.delete import *  # noqa
 from wagtail.admin.views.pages.edit import *  # noqa
 from wagtail.admin.views.pages.history import *  # noqa
 from wagtail.admin.views.pages.listing import *  # noqa
@@ -24,41 +24,6 @@ from wagtail.admin.views.pages.revisions import *  # noqa
 from wagtail.admin.views.pages.search import *  # noqa
 from wagtail.admin.views.pages.usage import *  # noqa
 from wagtail.admin.views.pages.workflow import *  # noqa
-
-
-def delete(request, page_id):
-    page = get_object_or_404(Page, id=page_id).specific
-    if not page.permissions_for_user(request.user).can_delete():
-        raise PermissionDenied
-
-    with transaction.atomic():
-        for fn in hooks.get_hooks('before_delete_page'):
-            result = fn(request, page)
-            if hasattr(result, 'status_code'):
-                return result
-
-        next_url = get_valid_next_url_from_request(request)
-
-        if request.method == 'POST':
-            parent_id = page.get_parent().id
-            page.delete(user=request.user)
-
-            messages.success(request, _("Page '{0}' deleted.").format(page.get_admin_display_title()))
-
-            for fn in hooks.get_hooks('after_delete_page'):
-                result = fn(request, page)
-                if hasattr(result, 'status_code'):
-                    return result
-
-            if next_url:
-                return redirect(next_url)
-            return redirect('wagtailadmin_explore', parent_id)
-
-    return TemplateResponse(request, 'wagtailadmin/pages/confirm_delete.html', {
-        'page': page,
-        'descendant_count': page.get_descendant_count(),
-        'next': next_url,
-    })
 
 
 def unpublish(request, page_id):
